% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_model_hierarchy}
\alias{check_model_hierarchy}
\title{Function for checking model hierarchy}
\usage{
check_model_hierarchy(model)
}
\value{
model hierarchy
}
\description{
Function for checking model hierarchy
}
\keyword{internal}
