% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_stepAIC_results.R
\name{collect_stepAIC_results}
\alias{collect_stepAIC_results}
\title{Collect Results From MASS::stepAIC Across Sample Datasets}
\usage{
collect_stepAIC_results(
  pvec_sample,
  ps_response,
  pvec_ignoring_columns = NULL,
  pvec_fixeffect_columns = NULL,
  pb_trace = FALSE
)
}
\arguments{
\item{pvec_sample}{vector of paths to sample files}

\item{ps_response}{response variable}

\item{pvec_ignoring_columns}{columns to be ignored in sample data}

\item{pvec_fixeffect_columns}{columns to be treated as fixed effects}

\item{pb_trace}{flag to indicate whether stepAIC produced output}
}
\value{
tbl_result tibble with model selection results
}
\description{
Model selection is done with \code{MASS::stepAIC()} for a set of sample
datasets. The results are collected into a tibble and returned to the
caller.

#' @examples
\dontrun{
n_nr_sample <- 10
n_sample_size <- 30
# samples from mtcars and write samples to files in tmpdir
tmpdir <- tempdir()
cat(" * Sample directory: ", tmpdir, "\n")
vec_sample_files <- sapply(1:n_nr_sample,
                           function(x) paste0("mtcars_sample_", x, ".csv", collapse = ""),
                           USE.NAMES = FALSE)
vec_sample_path <- file.path(tmpdir, vec_sample_files)
set.seed(321)
for (i in 1:n_nr_sample){
  cat(" * Generate sample ", i, " ...\n")
  cur_sample <- dplyr::sample_n(mtcars, n_sample_size, replace = TRUE)
  readr::write_csv(cur_sample, path = vec_sample_path[i])
}
collect_stepAIC_results(pvec_sample = vec_sample_path, ps_response = "mpg")
}
}
\details{
---
title: "Collection of MASS::stepAIC Results"
date:  "2020-04-29"
#' ---
}
