% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvergingResultFiles.R
\name{convergingResults_backAIC}
\alias{convergingResults_backAIC}
\title{Converging result-files of step backward aic in one overall result-file}
\usage{
convergingResults_backAIC(
  psDirectory = psDirectory,
  psSubdirectories = psSubdirectories,
  psOutputFile
)
}
\description{
Converging result-files of step backward aic in one overall result-file
}
