% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvergingResultFiles.R
\name{convergingResults_perSample_backAIC}
\alias{convergingResults_perSample_backAIC}
\title{Converging result-files of step backward aic in one result-file per sample}
\usage{
convergingResults_perSample_backAIC(psDirectory, psOutputFile)
}
\description{
Converging result-files of step backward aic in one result-file per sample
}
