% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat}
\alias{dat}
\title{A data frame contains dependent variable and independent variables}
\format{
\subsection{\code{dat}}{

A data frame.
}
}
\usage{
dat
}
\description{
A data frame with seven columns. The independent variables are in the first six columns. The dependent variable is in the seventh column.
}
\keyword{datasets}
