% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneratorOutputBackwardAIC.R
\name{generate_output_backAIC}
\alias{generate_output_backAIC}
\title{Generate an output for result of step backward for carcass}
\usage{
generate_output_backAIC(
  psInputFile,
  psTrait,
  psOutputPath,
  psSampleID,
  psTraitComment
)
}
\description{
Generate an output for result of step backward for carcass
}
