% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glmdat}
\alias{glmdat}
\title{A data frame contains dependent variable and binary independent variables}
\format{
\subsection{\code{glmdat}}{

A data frame.
}
}
\usage{
glmdat
}
\description{
A data frame with seven columns. The independent variables are in the first six columns. The dependent variable is in the seventh column.
}
\keyword{datasets}
