% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelselection.R
\name{ModelSelect}
\alias{ModelSelect}
\title{Title: Model selection for linear models}
\usage{
ModelSelect(formula, data)
}
\arguments{
\item{formula}{an object of class "formula": a symbolic description of the model to be fitted.
A typical model has the form \code{response ~ terms} where response is the (numeric) \code{response} vector and terms is a series of terms which specifies a linear predictor for \code{response}.
A terms specification of the form \code{first + second} indicates all the terms in \code{first} together with all the terms in \code{second} with duplicates removed.
A specification of the form \code{first:second} indicates the set of terms obtained by taking the interactions of all terms in \code{first} with all terms in \code{second}.
The specification \code{first*second} indicates the cross of \code{first} and \code{second.} This is the same as \code{first + second + first:second}.}

\item{data}{an data frame containing the variables in the model.}
}
\value{
\code{ModelSelect} returns a list containing the following components:

\describe{
\item{\code{model}}{A data frame of candidate models' BIC and posterior probabilities, sorted by decreasing posterior probability}
\item{\code{variable}}{A data frame of candidate variables' posterior inclusion probabilities and decision}
\item{\code{response}}{the response variable}
\item{\code{predictor}}{All candidate predictor variables.}
}

The function \code{lm.best} is used to obtain the linear fitting to the best model by posterior probability or by controlling variables' posterior inclusion probabilities.
}
\description{
Description: use BIC to do model selection.
}
