% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_stepAIC_results.R
\name{read_sample}
\alias{read_sample}
\title{Read a Sample File}
\usage{
read_sample(ps_sample_path)
}
\arguments{
\item{ps_sample_path}{path to sample file}
}
\value{
tbl_sample tibble read from sample file
}
\description{
The function distinguishes between files with extension "csv" and "txt" or "out".
Sample files with extension "csv" are read with readr::read_csv and sample files
with extension "txt" or "out" are read with readr::read_delim. Sample files with
extensions other than "csv", "txt" or "out" are not read and produce an error.
}
\details{
--- Reading Sample files ---
}
