% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_ModelSelectionCarcass.R
\name{running_ms_carcass}
\alias{running_ms_carcass}
\title{Running model selection for carcass}
\usage{
running_ms_carcass(psInputFile, psOutputPath, psSampleID)
}
\description{
Running model selection for carcass
}
