% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelselection.R
\name{selection}
\alias{selection}
\title{Model selection}
\usage{
selection(startm, endm, direction = "forward", trace = 1)
}
\arguments{
\item{startm}{regression for starting model}

\item{endm}{regression for ending model}

\item{direction}{direction of selection("backward" or "forward")}

\item{trace}{whether to output detailed process(1=Yes,0=No)}
}
\value{
the detailed process of adding or deleting covariates and final model
}
\description{
This package is used for backward or forward selection for linear regression
}
\examples{
data=matrix(rnorm(700),100,7)
colnames(data)=c("y","x1","x2","x3","x4","x5","x6")
data=data.frame(data)
lm1=lm(y~x1,data=data)
lm2=lm(y~.,data=data)
lm3=lm(y~x1+x3+x4+x5+x6,data=data)
selection(lm1,lm2,direction="forward",trace=1)
selection(lm2,lm2,direction="backward",trace=1)
selection(lm1,lm1,direction="backward",trace=1)

}
